@@ -19,16 +19,16 @@ classifiers = [
     "Programming Language :: Python :: 3",
     "Programming Language :: Python :: 3.10",
     "Programming Language :: Python :: 3.11",
-    "Programming Language :: Python :: 3.9",
+    "Programming Language :: Python :: 3.12",
     "Topic :: Other/Nonlisted Topic",
     "Topic :: Scientific/Engineering",
 ]
-requires-python = ">=3.9"
+requires-python = ">=3.10"
 dependencies = [
     "PyYAML",
     "click",
     "custodian>=2024.4.18",
-    "emmet-core==0.82.2",
+    "emmet-core>=0.82.2",
     "jobflow>=0.1.11",
     "monty>=2024.7.30",
     "numpy",
@@ -55,8 +55,18 @@ forcefields = [
     "chgnet>=0.2.2",
     "mace-torch>=0.3.3",
     "matgl>=1.1.3",
-    "quippy-ase>=0.9.14",
+    # quippy-ase support for py3.12 tracked in https://github.com/libAtoms/QUIP/issues/645
+    "quippy-ase>=0.9.14; python_version < '3.12'",
     "sevenn>=0.9.3",
+    "torchdata<=0.7.1",                            # TODO: remove when issue fixed
+]
+ase = ["ase>=3.23.0"]
+# tblite py3.12 support tracked in https://github.com/tblite/tblite/issues/198
+ase-ext = ["tblite>=0.3.0; python_version < '3.12'"]
+openmm = [
+    "mdanalysis>=2.7.0",
+    "openmm-mdanalysis-reporter>=0.1.0",
+    "openmm>=8.1.0",
 ]
 docs = [
     "FireWorks==2.0.3",
@@ -81,32 +91,39 @@ tests = [
 strict = [
     "PyYAML==6.0.2",
     "ase==3.23.0",
-    "calorine==2.2.1",
     "cclib==1.8.1",
-    "chgnet==0.3.8",
     "click==8.1.7",
     "custodian==2024.6.24",
     "dscribe==2.1.1",
-    "emmet-core==0.82.2",
+    "emmet-core==0.84.2",
     "ijson==3.3.0",
     "jobflow==0.1.18",
     "lobsterpy==0.4.5",
-    "mace-torch>=0.3.3",
-    "matgl==1.1.3",
+    "mdanalysis==2.7.0",
     "monty==2024.7.30",
     "mp-api==0.42.2",
     "numpy",
+    "openmm-mdanalysis-reporter==0.1.0",
+    "openmm==8.1.1",
     "phonopy==2.27.0",
     "pydantic-settings==2.5.2",
-    "pydantic==2.9.1",
+    "pydantic==2.9.2",
     "pymatgen-analysis-defects==2024.7.19",
     "pymatgen==2024.6.10",
     "python-ulid==2.7.0",
-    "quippy-ase==0.9.14",
     "seekpath==2.1.0",
+    "tblite==0.3.0; python_version < '3.12'",
+    "typing-extensions==4.12.2",
+]
+strict-forcefields = [
+    "calorine==2.2.1",
+    "chgnet==0.3.8",
+    "mace-torch>=0.3.3",
+    "matgl==1.1.3",
+    "quippy-ase==0.9.14; python_version < '3.12'",
     "sevenn==0.9.3.post1",
     "torch==2.2.1",
-    "typing-extensions==4.12.2",
+    "torchdata==0.7.1",                            # TODO: remove when issue fixed
 ]
 
 [project.scripts]
@@ -160,7 +177,8 @@ exclude_lines = [
 ]
 
 [tool.ruff]
-target-version = "py39"
+target-version = "py310"
+output-format = "concise"
 
 [tool.ruff.lint]
 select = ["ALL"]
