@@ -20,6 +20,7 @@ from atomate2.common.jobs.utils import structure_to_conventional, structure_to_p
 
 if TYPE_CHECKING:
     from pathlib import Path
+    from typing import Literal
 
     from emmet.core.math import Matrix3D
     from pymatgen.core.structure import Structure
@@ -28,7 +29,7 @@ if TYPE_CHECKING:
     from atomate2.forcefields.jobs import ForceFieldRelaxMaker, ForceFieldStaticMaker
     from atomate2.vasp.jobs.base import BaseVaspMaker
 
-SUPPORTED_CODES = frozenset(("vasp", "aims", "forcefields"))
+SUPPORTED_CODES = frozenset(("vasp", "aims", "forcefields", "ase"))
 
 
 @dataclass
@@ -133,7 +134,7 @@ class BasePhononMaker(Maker, ABC):
     min_length: float | None = 20.0
     prefer_90_degrees: bool = True
     get_supercell_size_kwargs: dict = field(default_factory=dict)
-    use_symmetrized_structure: str | None = None
+    use_symmetrized_structure: Literal["primitive", "conventional"] | None = None
     bulk_relax_maker: ForceFieldRelaxMaker | BaseVaspMaker | BaseAimsMaker | None = None
     static_energy_maker: ForceFieldRelaxMaker | BaseVaspMaker | BaseAimsMaker | None = (
         None
