@@ -7,7 +7,7 @@ from dataclasses import dataclass, field
 from atomate2 import SETTINGS
 from atomate2.common.flows.gruneisen import BaseGruneisenMaker
 from atomate2.forcefields.flows.phonons import PhononMaker
-from atomate2.forcefields.jobs import CHGNetRelaxMaker, ForceFieldRelaxMaker
+from atomate2.forcefields.jobs import ForceFieldRelaxMaker
 
 
 @dataclass
@@ -63,12 +63,14 @@ class GruneisenMaker(BaseGruneisenMaker):
 
     name: str = "Gruneisen"
     bulk_relax_maker: ForceFieldRelaxMaker | None = field(
-        default_factory=lambda: CHGNetRelaxMaker(relax_kwargs={"fmax": 0.00001})
+        default_factory=lambda: ForceFieldRelaxMaker(
+            force_field_name="CHGNet", relax_kwargs={"fmax": 0.00001}
+        )
     )
     code: str = "forcefields"
     const_vol_relax_maker: ForceFieldRelaxMaker = field(
-        default_factory=lambda: CHGNetRelaxMaker(
-            relax_kwargs={"fmax": 0.00001}, relax_cell=False
+        default_factory=lambda: ForceFieldRelaxMaker(
+            force_field_name="CHGNet", relax_kwargs={"fmax": 0.00001}, relax_cell=False
         )
     )
     kpath_scheme: str = "seekpath"
