@@ -3,14 +3,11 @@
 from __future__ import annotations
 
 from dataclasses import dataclass, field
-from typing import TYPE_CHECKING, Literal
+from typing import Literal
 
 from atomate2.common.flows.qha import CommonQhaMaker
 from atomate2.forcefields.flows.phonons import PhononMaker
-from atomate2.forcefields.jobs import CHGNetRelaxMaker
-
-if TYPE_CHECKING:
-    from atomate2.forcefields.jobs import ForceFieldRelaxMaker
+from atomate2.forcefields.jobs import ForceFieldRelaxMaker
 
 
 @dataclass
@@ -56,11 +53,11 @@ class CHGNetQhaMaker(CommonQhaMaker):
 
     name: str = "CHGNet QHA Maker"
     initial_relax_maker: ForceFieldRelaxMaker | None = field(
-        default_factory=CHGNetRelaxMaker
+        default_factory=lambda: ForceFieldRelaxMaker(force_field_name="CHGNet")
     )
     eos_relax_maker: ForceFieldRelaxMaker | None = field(
-        default_factory=lambda: CHGNetRelaxMaker(
-            relax_cell=False, relax_kwargs={"fmax": 0.00001}
+        default_factory=lambda: ForceFieldRelaxMaker(
+            force_field_name="CHGNet", relax_cell=False, relax_kwargs={"fmax": 0.00001}
         )
     )
     phonon_maker: PhononMaker = field(
