@@ -80,7 +80,7 @@ class EosMaker(CommonEosMaker):
     initial_relax_maker : .Maker | None
         Maker to relax the input structure, defaults to None (no initial relaxation).
     eos_relax_maker : .Maker
-        Maker to relax deformationed structures for the EOS fit.
+        Maker to relax deformed structures for the EOS fit.
     static_maker : .Maker | None
         Maker to generate statics after each relaxation, defaults to None.
     strain : tuple[float]
@@ -148,7 +148,7 @@ class MPLegacyEosMaker(CommonEosMaker):
     initial_relax_maker : .Maker | None
         Maker to relax the input structure, defaults to None (no initial relaxation).
     eos_relax_maker : .Maker
-        Maker to relax deformationed structures for the EOS fit.
+        Maker to relax deformed structures for the EOS fit.
     static_maker : .Maker | None
         Maker to generate statics after each relaxation, defaults to None.
     strain : tuple[float]
@@ -225,7 +225,7 @@ class MPGGAEosMaker(CommonEosMaker):
     initial_relax_maker : .Maker | None
         Maker to relax the input structure, defaults to None (no initial relaxation).
     eos_relax_maker : .Maker
-        Maker to relax deformationed structures for the EOS fit.
+        Maker to relax deformed structures for the EOS fit.
     static_maker : .Maker | None
         Maker to generate statics after each relaxation, defaults to None.
     strain : tuple[float]
@@ -293,7 +293,7 @@ class MPMetaGGAEosMaker(CommonEosMaker):
     initial_relax_maker : .Maker | None
         Maker to relax the input structure, defaults to None (no initial relaxation).
     eos_relax_maker : .Maker
-        Maker to relax deformationed structures for the EOS fit.
+        Maker to relax deformed structures for the EOS fit.
     static_maker : .Maker | None
         Maker to generate statics after each relaxation, defaults to None.
     strain : tuple[float]
