@@ -6,7 +6,7 @@ import warnings
 from dataclasses import dataclass, field
 from pathlib import Path
 from shutil import which
-from typing import TYPE_CHECKING, Callable
+from typing import TYPE_CHECKING
 
 from emmet.core.tasks import TaskDoc
 from jobflow import Maker, Response, job
@@ -26,6 +26,8 @@ from atomate2.vasp.run import run_vasp, should_stop_children
 from atomate2.vasp.sets.base import VaspInputGenerator
 
 if TYPE_CHECKING:
+    from collections.abc import Callable
+
     from pymatgen.core import Structure
 
 
