@@ -154,7 +154,7 @@ def run_elph_displacements(
         "uuids": [],
         "dirs": [],
     }
-    for temp, structure in zip(temperatures, structures):
+    for temp, structure in zip(temperatures, structures, strict=True):
         # create the job
         elph_job = vasp_maker.make(structure, prev_dir=prev_dir)
         elph_job.append_name(f" T={temp}")
