@@ -7,7 +7,7 @@ from atomate2.common.jobs.phonons import PhononBSDOSDoc
 from atomate2.common.jobs.qha import PhononQHADoc, analyze_free_energy
 
 
-def test_analyze_free_energy(clean_dir, test_dir):
+def test_analyze_free_energy(tmp_dir, test_dir):
     # The following code and the test files have been adapted from Phonopy
     # Copyright (C) 2015 Atsushi Togo
     # All rights reserved.
@@ -54,7 +54,7 @@ def test_analyze_free_energy(clean_dir, test_dir):
             energies.append(float(e))
 
     phonon_docs = []
-    for index, energy, volume in zip(range(-5, 6), energies, volumes):
+    for index, energy, volume in zip(range(-5, 6), energies, volumes, strict=True):
         filename = f"{test_dir}/qha/thermal_properties.yaml-{index!s}"
         yaml = YAML()
         with open(filename) as f:
@@ -84,7 +84,7 @@ def test_analyze_free_energy(clean_dir, test_dir):
     assert isinstance(qha_doc, PhononQHADoc)
 
 
-def test_analyze_free_energy_small(clean_dir, test_dir):
+def test_analyze_free_energy_small(tmp_dir, test_dir):
     # The following code and the test files have been adapted from Phonopy
     # Copyright (C) 2015 Atsushi Togo
     # All rights reserved.
@@ -131,7 +131,7 @@ def test_analyze_free_energy_small(clean_dir, test_dir):
             energies.append(float(e))
 
     phonon_docs = []
-    for index, energy, volume in zip(range(-5, 6), energies, volumes):
+    for index, energy, volume in zip(range(-5, 6), energies, volumes, strict=True):
         filename = f"{test_dir}/qha/thermal_properties.yaml-{index!s}"
         yaml = YAML()
         with open(filename) as f:
