@@ -1,10 +1,14 @@
 import pytest
-import torch
 from jobflow import run_locally
 from monty.serialization import loadfn
 
 from atomate2.forcefields import MLFF
-from atomate2.forcefields.flows.eos import CHGNetEosMaker, M3GNetEosMaker, MACEEosMaker
+from atomate2.forcefields.flows.eos import (
+    CHGNetEosMaker,
+    ForceFieldEosMaker,
+    M3GNetEosMaker,
+    MACEEosMaker,
+)
 
 ff_maker_map = {
     MLFF.CHGNet.value: CHGNetEosMaker,
@@ -15,10 +19,7 @@ ff_maker_map = {
 
 @pytest.mark.parametrize("mlff", ff_maker_map)
 def test_ml_ff_eos_makers(mlff: str, si_structure, clean_dir, test_dir):
-    # MACE changes the default dtype, ensure consistent dtype here
-    torch.set_default_dtype(torch.float32)
-
-    job = ff_maker_map[mlff]().make(si_structure)
+    job = ForceFieldEosMaker.from_force_field_name(mlff).make(si_structure)
     job_to_uuid = {job.name: job.uuid for job in job.jobs}
     post_process_uuid = job_to_uuid[f"{mlff} EOS Maker postprocessing"]
     response = run_locally(job, ensure_success=True)
@@ -34,3 +35,6 @@ def test_ml_ff_eos_makers(mlff: str, si_structure, clean_dir, test_dir):
                 output["relax"][key][idx] == pytest.approx(value)
                 for idx, value in ref_data["relax"][key].items()
             )
+
+    with pytest.warns(FutureWarning):
+        ff_maker_map[mlff]()
