@@ -10,6 +10,8 @@ from pytest import approx, importorskip
 from atomate2.forcefields.jobs import (
     CHGNetRelaxMaker,
     CHGNetStaticMaker,
+    ForceFieldRelaxMaker,
+    ForceFieldStaticMaker,
     GAPRelaxMaker,
     GAPStaticMaker,
     M3GNetRelaxMaker,
@@ -24,11 +26,26 @@ from atomate2.forcefields.jobs import (
 from atomate2.forcefields.schemas import ForceFieldTaskDocument
 
 
-def test_chgnet_static_maker(si_structure):
-    task_doc_kwargs = {"ionic_step_data": ("structure", "energy")}
+def test_maker_initialization():
+    # test that makers can be initialized from str or value enum
 
+    from atomate2.forcefields import MLFF
+
+    for mlff in MLFF.__members__:
+        assert ForceFieldRelaxMaker(
+            force_field_name=MLFF(mlff)
+        ) == ForceFieldRelaxMaker(force_field_name=mlff)
+        assert ForceFieldRelaxMaker(
+            force_field_name=str(MLFF(mlff))
+        ) == ForceFieldRelaxMaker(force_field_name=mlff)
+
+
+def test_chgnet_static_maker(si_structure):
     # generate job
-    job = CHGNetStaticMaker(task_document_kwargs=task_doc_kwargs).make(si_structure)
+    job = ForceFieldStaticMaker(
+        force_field_name="CHGNet",
+        ionic_step_data=("structure", "energy"),
+    ).make(si_structure)
 
     # run the flow or job and ensure that it finished running successfully
     responses = run_locally(job, ensure_success=True)
@@ -42,6 +59,9 @@ def test_chgnet_static_maker(si_structure):
 
     assert output1.forcefield_version == get_imported_version("chgnet")
 
+    with pytest.warns(FutureWarning):
+        CHGNetStaticMaker()
+
 
 @pytest.mark.parametrize(
     "fix_symmetry, symprec", [(True, 1e-2), (False, 1e-2), (True, 1e-1)]
@@ -53,7 +73,8 @@ def test_chgnet_relax_maker_fix_symmetry(
 ):
     # translate one atom to break symmetry but stay below symprec threshold
     ba_ti_o3_structure.translate_sites(1, [symprec / 10.0, 0, 0])
-    job = CHGNetRelaxMaker(
+    job = ForceFieldRelaxMaker(
+        force_field_name="CHGNet",
         relax_kwargs={"fmax": 0.01},
         fix_symmetry=fix_symmetry,
         symprec=symprec,
@@ -81,7 +102,11 @@ def test_chgnet_relax_maker(si_structure: Structure, relax_cell: bool):
 
     max_step = 25
     # generate job
-    job = CHGNetRelaxMaker(steps=max_step, relax_cell=relax_cell).make(si_structure)
+    job = ForceFieldRelaxMaker(
+        force_field_name="CHGNet",
+        steps=max_step,
+        relax_cell=relax_cell,
+    ).make(si_structure)
 
     # run the flow or job and ensure that it finished running successfully
     responses = run_locally(job, ensure_success=True)
@@ -103,12 +128,16 @@ def test_chgnet_relax_maker(si_structure: Structure, relax_cell: bool):
     # check the force_field_task_doc attributes
     assert Path(responses[job.uuid][1].output.dir_name).exists()
 
+    with pytest.warns(FutureWarning):
+        CHGNetRelaxMaker()
+
 
 def test_m3gnet_static_maker(si_structure):
-    task_doc_kwargs = {"ionic_step_data": ("structure", "energy")}
-
     # generate job
-    job = M3GNetStaticMaker(task_document_kwargs=task_doc_kwargs).make(si_structure)
+    job = ForceFieldStaticMaker(
+        force_field_name="M3GNet",
+        ionic_step_data=("structure", "energy"),
+    ).make(si_structure)
 
     # run the flow or job and ensure that it finished running successfully
     responses = run_locally(job, ensure_success=True)
@@ -121,6 +150,9 @@ def test_m3gnet_static_maker(si_structure):
 
     assert output1.forcefield_version == get_imported_version("matgl")
 
+    with pytest.warns(FutureWarning):
+        M3GNetStaticMaker()
+
 
 def test_m3gnet_relax_maker(si_structure):
     # translate one atom to ensure a small number of relaxation steps are taken
@@ -128,7 +160,10 @@ def test_m3gnet_relax_maker(si_structure):
 
     # generate job
     max_step = 25
-    job = M3GNetRelaxMaker(steps=max_step).make(si_structure)
+    job = ForceFieldRelaxMaker(
+        force_field_name="M3GNet",
+        steps=max_step,
+    ).make(si_structure)
 
     # run the flow or job and ensure that it finished running successfully
     responses = run_locally(job, ensure_success=True)
@@ -139,6 +174,9 @@ def test_m3gnet_relax_maker(si_structure):
     assert output1.is_force_converged
     assert output1.output.energy == approx(-10.8, abs=0.2)
     assert output1.output.n_steps == 24
+
+    with pytest.warns(FutureWarning):
+        M3GNetRelaxMaker()
 
 
 mace_paths = pytest.mark.parametrize(
@@ -153,12 +191,12 @@ mace_paths = pytest.mark.parametrize(
 
 @mace_paths
 def test_mace_static_maker(si_structure: Structure, test_dir: Path, model):
-    task_doc_kwargs = {"ionic_step_data": ("structure", "energy")}
-
     # generate job
     # NOTE the test model is not trained on Si, so the energy is not accurate
-    job = MACEStaticMaker(
-        calculator_kwargs={"model": model}, task_document_kwargs=task_doc_kwargs
+    job = ForceFieldStaticMaker(
+        force_field_name="MACE",
+        ionic_step_data=("structure", "energy"),
+        calculator_kwargs={"model": model},
     ).make(si_structure)
 
     # run the flow or job and ensure that it finished running successfully
@@ -171,6 +209,9 @@ def test_mace_static_maker(si_structure: Structure, test_dir: Path, model):
     assert output1.output.n_steps == 1
     assert output1.forcefield_version == get_imported_version("mace-torch")
 
+    with pytest.warns(FutureWarning):
+        MACEStaticMaker()
+
 
 @pytest.mark.parametrize(
     "fix_symmetry, symprec", [(True, 1e-2), (False, 1e-2), (True, 1e-1)]
@@ -182,7 +223,8 @@ def test_mace_relax_maker_fix_symmetry(
 ):
     # translate one atom to break symmetry but stay below symprec threshold
     ba_ti_o3_structure.translate_sites(1, [symprec / 10.0, 0, 0])
-    job = MACERelaxMaker(
+    job = ForceFieldRelaxMaker(
+        force_field_name="MACE",
         relax_kwargs={"fmax": 0.02},
         fix_symmetry=fix_symmetry,
         symprec=symprec,
@@ -201,6 +243,9 @@ def test_mace_relax_maker_fix_symmetry(
         assert initial_space_group == final_space_group
     else:
         assert initial_space_group != final_space_group
+
+    with pytest.warns(FutureWarning):
+        MACERelaxMaker()
 
 
 @pytest.mark.parametrize(
@@ -228,7 +273,8 @@ def test_mace_relax_maker(
 
     # generate job
     # NOTE the test model is not trained on Si, so the energy is not accurate
-    job = MACERelaxMaker(
+    job = ForceFieldRelaxMaker(
+        force_field_name="MACE",
         calculator_kwargs={"model": model},
         steps=25,
         optimizer_kwargs={"optimizer": "BFGSLineSearch"},
@@ -274,16 +320,15 @@ def test_mace_relax_maker(
 def test_gap_static_maker(si_structure: Structure, test_dir):
     importorskip("quippy")
 
-    task_doc_kwargs = {"ionic_step_data": ("structure", "energy")}
-
     # generate job
     # Test files have been provided by @YuanbinLiu (University of Oxford)
-    job = GAPStaticMaker(
+    job = ForceFieldStaticMaker(
+        force_field_name="GAP",
+        ionic_step_data=("structure", "energy"),
         calculator_kwargs={
             "args_str": "IP GAP",
             "param_filename": str(test_dir / "forcefields" / "gap" / "gap_file.xml"),
         },
-        task_document_kwargs=task_doc_kwargs,
     ).make(si_structure)
 
     # run the flow or job and ensure that it finished running successfully
@@ -296,18 +341,59 @@ def test_gap_static_maker(si_structure: Structure, test_dir):
     assert output1.output.n_steps == 1
     assert output1.forcefield_version == get_imported_version("quippy-ase")
 
+    with pytest.warns(FutureWarning):
+        GAPStaticMaker()
+
+
+@pytest.mark.parametrize("relax_cell", [True, False])
+def test_gap_relax_maker(si_structure: Structure, test_dir: Path, relax_cell: bool):
+    importorskip("quippy")
+
+    # translate one atom to ensure a small number of relaxation steps are taken
+    si_structure.translate_sites(0, [0, 0, 0.1])
+
+    # generate job
+    # Test files have been provided by @YuanbinLiu (University of Oxford)
+    max_step = 25
+    job = ForceFieldRelaxMaker(
+        force_field_name="GAP",
+        calculator_kwargs={
+            "args_str": "IP GAP",
+            "param_filename": str(test_dir / "forcefields" / "gap" / "gap_file.xml"),
+        },
+        steps=max_step,
+        relax_cell=relax_cell,
+    ).make(si_structure)
+
+    # run the flow or job and ensure that it finished running successfully
+    responses = run_locally(job, ensure_success=True)
+
+    # validating the outputs of the job
+    output1 = responses[job.uuid][1].output
+    assert isinstance(output1, ForceFieldTaskDocument)
+    if relax_cell:
+        assert not output1.is_force_converged
+        assert output1.output.energy == approx(-13.08492, rel=1e-2)
+        assert output1.output.n_steps == max_step + 2
+    else:
+        assert output1.is_force_converged
+        assert output1.output.energy == approx(-10.8523, rel=1e-4)
+        assert output1.output.n_steps == 17
+
+    with pytest.warns(FutureWarning):
+        GAPRelaxMaker()
+
 
 def test_nep_static_maker(al2_au_structure: Structure, test_dir: Path):
-    task_doc_kwargs = {"ionic_step_data": ("structure", "energy")}
-
     # NOTE: The test NEP model is specifically trained on 16 elemental metals
     # thus a new Al2Au structure is added.
     # The NEP model used for the tests is licensed under a
     # [CC BY 4.0](https://creativecommons.org/licenses/by/4.0/legalcode)
     # and downloaded from https://doi.org/10.5281/zenodo.10081677
     # Also cite the original work if you use this specific model : https://arxiv.org/abs/2311.04732
-    job = NEPStaticMaker(
-        task_document_kwargs=task_doc_kwargs,
+    job = ForceFieldStaticMaker(
+        force_field_name="NEP",
+        ionic_step_data=("structure", "energy"),
         calculator_kwargs={
             "model_filename": test_dir / "forcefields" / "nep" / "nep.txt"
         },
@@ -321,6 +407,9 @@ def test_nep_static_maker(al2_au_structure: Structure, test_dir: Path):
     assert isinstance(output1, ForceFieldTaskDocument)
     assert output1.output.energy == approx(-47.65972, rel=1e-4)
     assert output1.output.n_steps == 1
+
+    with pytest.warns(FutureWarning):
+        NEPStaticMaker()
 
 
 @pytest.mark.parametrize(
@@ -341,7 +430,8 @@ def test_nep_relax_maker(
     # Also cite the original work if you use this specific model : https://arxiv.org/abs/2311.04732
 
     # generate job
-    job = NEPRelaxMaker(
+    job = ForceFieldRelaxMaker(
+        force_field_name="NEP",
         steps=25,
         optimizer_kwargs={"optimizer": "BFGSLineSearch"},
         relax_cell=relax_cell,
@@ -369,15 +459,18 @@ def test_nep_relax_maker(
     final_spg_num = output1.output.structure.get_space_group_info()[1]
     assert final_spg_num == 225
 
+    with pytest.warns(FutureWarning):
+        NEPRelaxMaker()
+
 
 def test_nequip_static_maker(sr_ti_o3_structure: Structure, test_dir: Path):
     importorskip("nequip")
-    task_doc_kwargs = {"ionic_step_data": ("structure", "energy")}
 
     # generate job
     # NOTE the test model is not trained on Si, so the energy is not accurate
-    job = NequipStaticMaker(
-        task_document_kwargs=task_doc_kwargs,
+    job = ForceFieldStaticMaker(
+        force_field_name="Nequip",
+        ionic_step_data=("structure", "energy"),
         calculator_kwargs={
             "model_path": test_dir / "forcefields" / "nequip" / "nequip_ff_sr_ti_o3.pth"
         },
@@ -392,6 +485,9 @@ def test_nequip_static_maker(sr_ti_o3_structure: Structure, test_dir: Path):
     assert output1.output.energy == approx(-44.40017, rel=1e-4)
     assert output1.output.n_steps == 1
     assert output1.forcefield_version == get_imported_version("nequip")
+
+    with pytest.warns(FutureWarning):
+        NequipStaticMaker()
 
 
 @pytest.mark.parametrize(
@@ -408,7 +504,8 @@ def test_nequip_relax_maker(
     # translate one atom to ensure a small number of relaxation steps are taken
     sr_ti_o3_structure.translate_sites(0, [0, 0, 0.2])
     # generate job
-    job = NequipRelaxMaker(
+    job = ForceFieldRelaxMaker(
+        force_field_name="Nequip",
         steps=25,
         optimizer_kwargs={"optimizer": "BFGSLineSearch"},
         relax_cell=relax_cell,
@@ -436,37 +533,5 @@ def test_nequip_relax_maker(
     final_spg_num = output1.output.structure.get_space_group_info()[1]
     assert final_spg_num == 99
 
-
-@pytest.mark.parametrize("relax_cell", [True, False])
-def test_gap_relax_maker(si_structure: Structure, test_dir: Path, relax_cell: bool):
-    importorskip("quippy")
-
-    # translate one atom to ensure a small number of relaxation steps are taken
-    si_structure.translate_sites(0, [0, 0, 0.1])
-
-    # generate job
-    # Test files have been provided by @YuanbinLiu (University of Oxford)
-    max_step = 25
-    job = GAPRelaxMaker(
-        calculator_kwargs={
-            "args_str": "IP GAP",
-            "param_filename": str(test_dir / "forcefields" / "gap" / "gap_file.xml"),
-        },
-        steps=max_step,
-        relax_cell=relax_cell,
-    ).make(si_structure)
-
-    # run the flow or job and ensure that it finished running successfully
-    responses = run_locally(job, ensure_success=True)
-
-    # validating the outputs of the job
-    output1 = responses[job.uuid][1].output
-    assert isinstance(output1, ForceFieldTaskDocument)
-    if relax_cell:
-        assert not output1.is_force_converged
-        assert output1.output.energy == approx(-13.08492, rel=1e-2)
-        assert output1.output.n_steps == max_step + 2
-    else:
-        assert output1.is_force_converged
-        assert output1.output.energy == approx(-10.8523, rel=1e-4)
-        assert output1.output.n_steps == 17
+    with pytest.warns(FutureWarning):
+        NequipRelaxMaker()
