@@ -1,5 +1,6 @@
 """Tests for forcefield MD flows."""
 
+import sys
 from pathlib import Path
 
 import numpy as np
@@ -14,6 +15,7 @@ from pymatgen.core import Structure
 
 from atomate2.forcefields.md import (
     CHGNetMDMaker,
+    ForceFieldMDMaker,
     GAPMDMaker,
     M3GNetMDMaker,
     MACEMDMaker,
@@ -31,6 +33,20 @@ name_to_maker = {
 }
 
 
+def test_maker_initialization():
+    # test that makers can be initialized from str or value enum
+
+    from atomate2.forcefields import MLFF
+
+    for mlff in MLFF.__members__:
+        assert ForceFieldMDMaker(force_field_name=MLFF(mlff)) == ForceFieldMDMaker(
+            force_field_name=mlff
+        )
+        assert ForceFieldMDMaker(force_field_name=str(MLFF(mlff))) == ForceFieldMDMaker(
+            force_field_name=mlff
+        )
+
+
 @pytest.mark.parametrize(
     "ff_name",
     ["CHGNet", "M3GNet", "MACE", "GAP", "NEP", "Nequip"],
@@ -38,6 +54,11 @@ name_to_maker = {
 def test_ml_ff_md_maker(
     ff_name, si_structure, sr_ti_o3_structure, al2_au_structure, test_dir, clean_dir
 ):
+    if ff_name == "GAP" and sys.version_info >= (3, 12):
+        pytest.skip(
+            "GAP model not compatible with Python 3.12, waiting on https://github.com/libAtoms/QUIP/issues/645"
+        )
+
     n_steps = 5
 
     ref_energies_per_atom = {
@@ -78,11 +99,13 @@ def test_ml_ff_md_maker(
 
     structure = unit_cell_structure.to_conventional() * (2, 2, 2)
 
-    job = name_to_maker[ff_name](
+    job = ForceFieldMDMaker(
+        force_field_name=ff_name,
         n_steps=n_steps,
         traj_file="md_traj.json.gz",
         traj_file_fmt="pmg",
-        task_document_kwargs={"store_trajectory": "partial"},
+        store_trajectory="partial",
+        ionic_step_data=("energy", "forces", "stress", "mol_or_struct"),
         calculator_kwargs=calculator_kwargs,
     ).make(structure)
     response = run_locally(job, ensure_success=True)
@@ -101,18 +124,22 @@ def test_ml_ff_md_maker(
     # Check that the ionic steps have the expected physical properties
     assert all(
         key in step.model_dump()
-        for key in ("energy", "forces", "stress", "structure")
+        for key in ("energy", "forces", "stress", "mol_or_struct", "structure")
         for step in task_doc.output.ionic_steps
     )
 
     # Check that the trajectory has expected physical properties
     assert task_doc.included_objects == ["trajectory"]
-    assert len(task_doc.forcefield_objects["trajectory"]) == n_steps + 1
+    assert len(task_doc.objects["trajectory"]) == n_steps + 1
+    assert task_doc.objects == task_doc.forcefield_objects  # test legacy alias
     assert all(
         key in step
         for key in ("energy", "forces", "stress", "velocities", "temperature")
-        for step in task_doc.forcefield_objects["trajectory"].frame_properties
+        for step in task_doc.objects["trajectory"].frame_properties
     )
+
+    with pytest.warns(FutureWarning):
+        name_to_maker[ff_name]()
 
 
 @pytest.mark.parametrize("traj_file", ["trajectory.json.gz", "atoms.traj"])
@@ -130,7 +157,8 @@ def test_traj_file(traj_file, si_structure, clean_dir, ff_name="CHGNet"):
         traj_file_loader = Trajectory
 
     structure = si_structure.to_conventional() * (2, 2, 2)
-    job = name_to_maker[ff_name](
+    job = ForceFieldMDMaker(
+        force_field_name=ff_name,
         n_steps=n_steps,
         traj_file=traj_file,
         traj_file_fmt=traj_file_fmt,
@@ -146,9 +174,7 @@ def test_traj_file(traj_file, si_structure, clean_dir, ff_name="CHGNet"):
         assert all(
             np.all(
                 traj_from_file.frame_properties[idx][key]
-                == task_doc.forcefield_objects["trajectory"]
-                .frame_properties[idx]
-                .get(key)
+                == task_doc.objects["trajectory"].frame_properties[idx].get(key)
             )
             for key in ("energy", "temperature", "forces", "velocities")
             for idx in range(n_steps + 1)
@@ -166,9 +192,7 @@ def test_traj_file(traj_file, si_structure, clean_dir, ff_name="CHGNet"):
         assert all(
             np.all(
                 traj_as_dict[idx - 1][key]
-                == task_doc.forcefield_objects["trajectory"]
-                .frame_properties[idx]
-                .get(key)
+                == task_doc.objects["trajectory"].frame_properties[idx].get(key)
             )
             for key in ("energy", "temperature", "forces", "velocities")
             for idx in range(1, n_steps + 1)
@@ -188,32 +212,36 @@ def test_nve_and_dynamics_obj(si_structure: Structure, test_dir: Path):
         elif key == "from_dyn":
             dyn = VelocityVerlet
 
-        job = CHGNetMDMaker(
+        job = ForceFieldMDMaker(
+            force_field_name="CHGNet",
             ensemble="nve",
             dynamics=dyn,
             n_steps=50,
             traj_file=None,
+            ionic_step_data=("energy", "forces", "stress", "structure"),
         ).make(si_structure)
 
         response = run_locally(job, ensure_success=True)
-        output[key] = response[next(iter(response))][1].output
+        output[key] = response[job.uuid][1].output
 
     # check that energy and volume are constants
-    assert output["from_str"].output.energy == pytest.approx(-10.6, abs=0.1)
+    ref_toten = -10.6
+    assert output["from_str"].output.energy == pytest.approx(ref_toten, abs=0.1)
     assert output["from_str"].output.structure.volume == pytest.approx(
         output["from_str"].input.structure.volume
     )
+
     assert all(
-        step.energy == pytest.approx(-10.6, abs=0.1)
+        step.energy == pytest.approx(ref_toten, abs=0.1)
         for step in output["from_str"].output.ionic_steps
     )
-
     # ensure that output is consistent if molecular dynamics object is specified
     # as str or as MolecularDynamics object
     for attr in ("energy", "forces", "stress", "structure"):
         vals = {
             key: getattr(output[key].output, attr) for key in ("from_str", "from_dyn")
         }
+
         if isinstance(vals["from_str"], float):
             assert vals["from_str"] == pytest.approx(vals["from_dyn"])
         elif isinstance(vals["from_str"], Structure):
@@ -233,7 +261,8 @@ def test_temp_schedule(ff_name, si_structure, clean_dir):
 
     structure = si_structure.to_conventional() * (2, 2, 2)
 
-    job = name_to_maker[ff_name](
+    job = ForceFieldMDMaker(
+        force_field_name=ff_name,
         n_steps=n_steps,
         traj_file=None,
         dynamics="nose-hoover",
@@ -244,8 +273,7 @@ def test_temp_schedule(ff_name, si_structure, clean_dir):
     task_doc = response[next(iter(response))][1].output
 
     temp_history = [
-        step["temperature"]
-        for step in task_doc.forcefield_objects["trajectory"].frame_properties
+        step["temperature"] for step in task_doc.objects["trajectory"].frame_properties
     ]
 
     assert temp_history[-1] > temp_schedule[0]
@@ -258,7 +286,8 @@ def test_press_schedule(ff_name, si_structure, clean_dir):
 
     structure = si_structure.to_conventional() * (3, 3, 3)
 
-    job = name_to_maker[ff_name](
+    job = ForceFieldMDMaker(
+        force_field_name=ff_name,
         ensemble="npt",
         n_steps=n_steps,
         traj_file="md_traj.json.gz",
@@ -271,7 +300,6 @@ def test_press_schedule(ff_name, si_structure, clean_dir):
         },
     ).make(structure)
     run_locally(job, ensure_success=True)
-    # task_doc = response[next(iter(response))][1].output
 
     traj_from_file = loadfn("md_traj.json.gz")
 
