@@ -27,7 +27,7 @@ def test_phonon_maker_initialization_with_all_mlff(
 
     chk_pt_dir = test_dir / "forcefields"
     for mlff in MLFF:
-        if mlff == MLFF.GAP:
+        if mlff in {MLFF.GAP, MLFF.Forcefield}:
             continue  # TODO fix GAP, currently fails with RuntimeError, see
             # https://github.com/materialsproject/atomate2/pull/918#issuecomment-2253659694
         calc_kwargs = {
