@@ -3,10 +3,11 @@ from __future__ import annotations
 import logging
 import shutil
 from pathlib import Path
-from typing import TYPE_CHECKING, Any, Callable, Final, Literal
+from typing import TYPE_CHECKING, Any, Final, Literal
 
 import pytest
 from jobflow import CURRENT_JOB
+from pymatgen.core import Molecule
 from pymatgen.io.qchem.inputs import QCInput
 from pytest import MonkeyPatch
 
@@ -16,7 +17,7 @@ import atomate2.qchem.run
 from atomate2.qchem.sets.base import QCInputGenerator
 
 if TYPE_CHECKING:
-    from collections.abc import Generator, Sequence
+    from collections.abc import Callable, Generator, Sequence
 
 
 logger = logging.getLogger("atomate2")
@@ -24,6 +25,18 @@ logger = logging.getLogger("atomate2")
 _QFILES: Final = "mol.qin.gz"
 _REF_PATHS: dict[str, str | Path] = {}
 _FAKE_RUN_QCHEM_KWARGS: dict[str, dict] = {}
+
+
+@pytest.fixture
+def h2o_molecule():
+    return Molecule(
+        coords=[
+            [0.0000, 0.0000, 0.12124],
+            [-0.78304, -0.00000, -0.48495],
+            [0.78304, -0.00000, -0.48495],
+        ],
+        species=["O", "H", "H"],
+    )
 
 
 @pytest.fixture(scope="session")
@@ -169,6 +182,9 @@ def check_qin(
         user_qin_path = script_directory / "opt.qin.gz"
     elif job_name == "water_frequency":
         user_qin_path = script_directory / "freq.qin.gz"
+    else:
+        user_qin_path = Path("mol.qin")
+
     user_qin = QCInput.from_file(user_qin_path)
 
     keys_to_check = (
